+++ functions/faq.js
@@ -9,9 +9,8 @@ const faq = async (msg, params, flags) => {
 			.setColor('ffa500')
 			.setTitle(data.question)
 			.setDescription(format(data.answer))
-			.setFooter("Data pulled from Official rules v1.3 May, 2019")
-			.setURL("https://images-cdn.fantasyflightgames.com/filer_public/f9/a2/f9a28865-cd96-4f36-97e5-0f8ea75288c0/keyforge_rulebook_v9-compressed.pdf");
-
+			.setFooter("Data pulled from Official rules v1.4 Sept, 2019")
+			.setURL("https://images-cdn.fantasyflightgames.com/filer_public/88/71/8871df4e-5647-4a22-a8cc-4e6b0a46a15c/keyforge_rulebook_v10-compressed.pdf");
 		main.sendMessage(msg, {embed});
 	}
 };