+++ functions/help.js
@@ -36,6 +36,7 @@ const help = (msg, params) => {
 				.addField(`${prefix}invite`, 'Get the invite link.')
 				.addField(`${prefix}version`, 'Get the version number of current bot.')
 				.addField(`${prefix}faq searchTerm`, 'Get any FAQ containing specific term.')
+				.addField(`${prefix}Timing searchTerm`, 'Get any the timing Chart containing specific term.')
 				.addField(`[card name 1] [card name 2] (-en/-es/-it/-de/-fr/-pl/-pt/-th/-zh)`, 'Searches and displays card as an single image (up to 5) [] required. Add -lang tag to specify language.')
 				.addField('Bot Information', '[SkyJedi\'s Bot Emporium](https://discord.gg/G8au6FH)')
 				.addField('Other Info', ' [KeyForge Discord Server](https://discordapp.com/invite/PcTGhr9) \n[KeyForge](https://www.fantasyflightgames.com/en/products/keyforge/?powered_by=archonMatrixDiscord) by Fantasy Flight Game');